freifunk-common: neigh.sh: fix for new olsrd 0.9.5
Starting with version 0.9.5 the new olsr-deamon behaves
more standard-compliant and outputs correct HTTP-headers.
This would confuse the script and abort parsing.

Fix that by only trying to parse the JSON-value and
just ignore the other stuff. After this, we can see neighbours again.

Thanks to FreifunkUFO <ufo@rund.freifunk.net> for spotting this.