Add support for 802.11r fast transition to wireless config
Adds support for 802.11r fast transition to luci.
Corrects the deletion on nasid in issue #844

The R0KH & R1KH options are not perfect as it would be better
to separate out each option and do some sanity checking on
the values, but that is beyond my capabilities.

Signed-off-by: Warren Linton warren@linton.id.au

(squashed from three original commits in #994. commit message reworded)