luci-app-statistics: cpufreq: changes in data structure
Collectd 5.9.0 changed the data structure of the cpufreq plugin:
CPU cores are now handled as separate plugin instances.

There are also new data items per core:
 * time spent at each frequency
 * amount of frequency transitions

Enable these new data items, but initially hide them behind
a new config option "ExtraItems" (default: disabled), as
the amount of graphs in multi-core systems could be rather large.

Note that the frequencies are not (yet) sorted, so the
information value of the time-spent graph is semi-random.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>