luci.mk: Only install a default post-install script if there is none defined
If a postinst script is defined in the package Makefile, it will
not be included because it will be replaced by the default postinst
script. To solve the problem, we now check, if we have a postinst script.
And if so, it will be installed instead of the default postinst script.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>