luci-app-attendedsysupgrade: Send target to server
Currently the running devices `board_name` is mapped to it's building
`profile` name. In case of inconsistent `board_names` like the x86/64
target, it is hard to create a mapping that translates
`lenovo-20q1s02m00` to x86/64 and the generic profile.

To handle such cases, send the `target` per default.

Also `model` is no longer needed as no heuristics are applied anymore to
find the relation between model and profile.

Signed-off-by: Paul Spooren <mail@aparcar.org>