luci-mod-network: fix unbalanced and confusing markup in DHCP translation
The string `<%s>...<%s>` is invalid HTML and will confuse translation software
like poedit as well as human translators. It is not clear that `<%s>...<%s>`
is supposed to interpolate into both a start and end tag and `<%s>` might be
mistaken as a quoted form of `%s` like `«%s»` or `〈%s〉`.

Replace the string with `<a href="%s">...</a>` to make it immediately clear
that the acronym is surrounded by an HTML anchor, only pass the destination
URL via string formatting to avoid mistranslations there.

Fixes: 3bfcf99a68 ("luci-mod-network: fix unbalanced and confusing markup in DHCP translation")
Signed-off-by: Jo-Philipp Wich <jo@mein.io>