luci-theme-openwrt-2020: fix font issue when mixing latin and non-latin symbols
The GalanoGrotesque font used by the openwrt2020 theme does not support non-
latin symbols.

If latin and non-latin symbols are used together in one line it looks strange
because for latin symbols GalanoGrotesque is used but for non-latin symbols a
fallback font is used (sans-serif).

This patch changes default font to Helvetica for the "bg", "ru", "uk", "el"
and "he" locales.

Original patch was written by Jo-Philipp Wich

Fixes: #5580
Signed-off-by: Sergey V. Lobanov <sergey@lobanov.in>
[reword commit message, drop "de" from exception list]
Signed-off-by: Jo-Philipp Wich <jo@mein.io>