luci-material-theme: fix localizated left menu
The latest fix to fix the logout worked, but it was not applied to all
the resolutions, and I have observed the same problem exists for the
dasboard optional component.
Looking at the code, it seems only the components without submenu, have the
data-title element, so this fix modifies the style for all of them, not
relying on the position or text content.

Signed-off-by: Miguel Angel Mulero Martinez <migmul@gmail.com>