luci: remove installation of attendedsysupgrade
The package is only useful when running an official build. The
ImageBuilders used for this service can not contain any special compiler
flags etc. used by custom builds.

Instead of installing the package whenever `luci` is installed, it
should be installed via the buildbots itself, like it's currently done
for `luci`.

Signed-off-by: Paul Spooren <mail@aparcar.org>