luci-app-acme: Split Challenge Validation tab into two
The validation_method is now set to webroot by default.
This is the most used option.

The webroot option is now empty by default.
Into the description added that by default will be used /var/run/acme/challenge/
Webservers should serve the folder under /.well-known/acme-challenge/ url.
The folder is automatically created by the acmesh service on renewal.

The Challenge Validation Tab is split to Webroot and DNS and the validation method moved to general tab.
So ideally a user won't see the webroot folder option from other tab.
And a user can basically enable the webroot just in the general tab without leaving it.
The DNS validation needs too many options to it needs for own tab.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>