luci-mod-network: improve descriptions for DSL annex option values
Spell out the names of the DSL standards to make the options easier to
understand. For modems supporting VDSL, also mention clearly which of
the options actually set the bits for VDSL2.

As the items now start with the name of an ADSL standard, use that for
ordering. The generic variants including VDSL2 support stay at the top,
as they should be a good choice in almost all cases.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>