luci-app-acme: Reuse translations for removed state_dir option
In the commit 7a005f11261e1b6d01fb8e5a26b9c9862efd4e68 was removed the state_dir.
But translations were kept and they still have a value.
We should say to a user where the certs are stored now, and we can reuse the old translations.
They need to be fixed a little bit and already fixed for Russian, Ukrainian and Bulgarian languages.

See #6273

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>