luci-base: fix license
During the update of luasrcdiet with commit 5d0e02ff8c5495a76640536dd64734e7b607da7e
the license was incorrectly changed from 'Apache-2.0' to 'MIT'.

As the name implies, this is the basic package that must be installed on the
target for the LuCI to work at all. So changing the license from 'Apache-2.0'
to 'MIT' should not have happend, because this information belongs to the
package 'luci-base' and to the 'luasrcdiet'.

Therefore, with commit 0a6eb5a33167754710b1e6067f494e52257f1acd the
'luasrcdiet' was moved to its own package. However, it was forgotten to adapt
the licence to 'Apache-2.0' again.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>