luci-mod-network: fix look and feel for DNSMASQ IP-Sets generation
On the configuration page of '/admin/network/dhcp' on the IP sets tab,
it is possible to make changes directly in the table. However, this is
not displayed correctly when using the theme 'luci-theme-material' if the
options can be changed directly in this view.

The problem is solved by not editing the configuration in the overview page
but in the overlay page.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>