luci-base: use json for ucitrack instead of uci
The 'ucitrack' configuration under '/etc/config' defines which service must
be restarted on a LuCI change. This uci file defines how all this works
together. Additionally there are 'uci-defaults' scripts in the various
applications that add additional configurations parameters to this ucitrack
file if a service must be reloaded/restarted on a LuCI change.

There are some problems with the current implementation:

* The uci should be used to configure the system and not for this kind of
  reload/restart dependency handling on LuCI changes.
* On a system update with configuration restore of the 'ucitrack' file
  the new behavior on LuCI reload/restart could not take into account
  because the new file is *not* used.

This commit converts the handling from uci to json.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>