luci-app-firewall: SNAT display fixes
The implied logic in the functions to build display text is a bit of
a minefield.

Now, if family=4+6 is selected, display 4+6.

This part is a follow-up to: 909d2356c3f8a22348a2da5296f9e43b344c023c

Previously it would display only IPv6.

Now, if family=auto i.e. '' is selected then display 4 only.
fw4 internally treats no family as IPv4 only, meaning that IPv6 was not
SNATed. (This treatment is 'incorrect' but because it has always been
this way, this behaviour is retained for backwards compatibility, and
user expectations.)

Previous logic was incorrect (bug) and would display:

Forwarded IPv4 and IPv6

misleading the user.

Signed-off-by: Paul Donald <newtwen+github@gmail.com>