luci-app-lxc: remove bypassing GPG validation
Since https://github.com/lxc/lxc/commit/58520263041b6864cadad96278848f9b8ce78ee9
which is part of LXC 5.0, not being able to bypass GPG validation anymore has
been enforced by upstream in lxc-download. When LXC was bumped to 5.0.1 in
https://github.com/openwrt/packages/commit/d957a2293b2a21b1edca1aa92e141bad8292251a
this broke passing the --no-validate option to lxc-download, which resulted
in getting the output of the --help command line option as the RPC output of
lxc-download as it doesn't understand the --no-validate option. This in turn
broke the string parsing for compatible distributions and their versions.

To solve this, this commit removes the --no-validate option entirely as it has
been removed by upstream.

Signed-off-by: Daniel Nilsson <daniel.nilsson94@outlook.com>