luci-app-keepalived: fix nopreempt value handling
The 'nopreempt' option is marked with the prefix 'no_value_' in the
keepalived init script. This means that if the value is not set in the
keepalived configuration, it must be removed from the uci configuration.

To get this behavior, the rmempty flag in the LuCI must be removed.

fixes #7114

Signed-off-by: Florian Eckert <fe@dev.tdt.de>