luci-app-acme: Remove use_acme_server option since it's ignored by acme
In the fe48f1de0a1720b5f1281484edba9cddcdfe9266 was added use_acme_server flag and acme_server URL.
The problem is that the acme uses the only acme_server option.
A user may think that once the flag is turned off it will be ignored, but it's not.

To avoid confusion, it's easier just to remove the flag.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>