luci-base: form.js: decode HTML entities in AbstractElement.stripTags()
This commit fixes a problem with HTML entities which were visible in their
encoded form in the mobile view. This happened for example when displaying
a GridSection with a Value option containing "&nbsp;" in the title.
Without this change only HTML entities in titles that also contains tags
are decoded before they are stored in data-title attributes.

Signed-off-by: Mikael Magnusson <mikma@users.sourceforge.net>