luci: Make password reveal work with passsword managers
Password managers (like LastPass etc) tend to add additional elements
into the DOM for their own context menus. If this happens between the
hide/reveal button and the password input, then the logic to reveal the
password breaks.

This change updates the onclick handler to look for the first `<input>`
element with the class `password-input` that is under the parent of the
toggle button, and then to toggle the password/text type on that
element.

This change deliberately only updates the main ui.js file, not any
application files.

Signed-off-by: Andrew Dodd <andrew.john.dodd@gmail.com>