modules: dhcp: improve wording for DHCP reservations
The use of "static" regarding DHCP reservations can be misleading,
as static usually implies a setting on the host device, where what
we're actually doing is reserving the address on the server.

Change the button label from "Set Static" to "Reserve IP" to more
clearly indicate the action to be taken.  Add some tip text to make
it even more clear.

Fixes: https://github.com/openwrt/luci/issues/7801
Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>