luci-base: give TextValue form elements separate readonly and disabled
textarea widgets have more value as readonly and not disabled, than
just disabled: you can scroll through them when they're readonly and
not disabled.

So give the form element separate readonly and disabled properties, so
textarea elements defined as readonly with a large amount of content
still cannot be modified, but at least the whole content can be viewed
(because the textarea scrollbars still function).

Signed-off-by: Paul Donald <newtwen+github@gmail.com>