luci-base: extend FileUpload; create DirectoryPicker
FileUpload was extended to accommodate the new features, since
it has nearly everything already. Create a DirectoryPicker
convenience wrapper.

Additions are:
-directory creation (dialogue); set directory_create to true
-directory select mode (instead of file); set directory_select to true

Also fix a bug in the breadcrumb generation which produced:

/foo » » bar

for /foo/bar if root_directory is not '/', and another bug that
merged links together when navigating upward again using the
breadcrumbs.

Signed-off-by: Paul Donald <newtwen+github@gmail.com>