luci-base: jsdoc Slider -> RangeSlider and fixes for RangeSlider
The usecalc property suffers from recursive calculation; that is, its
output becomes its input at the next save. It is not known in advance
whether a stored value is one that was calculated or not. So this part
was removed. The getCalculatedValue() function is retained should it be
desirable to get this value.

The 'optional' property was removed since it didn't do anything.
The 'validate' property is now correctly bound.

Signed-off-by: Paul Donald <newtwen+github@gmail.com>