luci-mod-network,-status: show DHCPv6 lease IAID
The IAID is important information as it allows the user to know which
interface on the client device a given DHCPv6 lease corresponds to.
odhcpd already exposes this information (e.g. via "ubus call dhcp
ipv6leases"), but it would be good to have access to the same
information via the web interface, especially since odhcpd does take the
IAID into account when allocating addresses.

Signed-off-by: David Härdeman <david@hardeman.nu>