luci-mod-network: organize "devices" tab in dhcp/dns views
Sort the options in the "devices" tab so that they are all grouped together in
the order they appear in the UI and delete ones that are specific to DNS from
dhcp.js and vice versa.

Note that this means that the "nonwildcard", "interface" and "notinterface"
options are present in both views, since they control the interfaces that
dnsmasq listens to both for DNS and DHCP.

We might consider adding a help text blurb explaining that, but I've
intentionally not changed any code in these patches.

Signed-off-by: David Härdeman <david@hardeman.nu>