luci-mod-network: add odhcpd cfg in DHCP server setup
The "Set up DHCP Server" button is meant to setup the basics for a DHCP server
on the given interface.  Add odhcpd setting "dhcpv4" as well. It is a no-op in
case odhcpd isn't setup to take care of DHCPv4 as well, but it makes it easier
to support either dnsmasq or odhpcd as the DHCPv4 server.

One might argue that the button should also set dhcpv6=server and ra=server, so
that it doesn't only setup a DHCPv4 server, but also prepares a DHCPv6 server,
but that's for another PR :)

Signed-off-by: David Härdeman <david@hardeman.nu>